aider: Fix accessing author's name by concatenating first_name and last_name.
# Aider chat conversation:

USER: I ran this command:

 python main.py

And got this output:

Traceback (most recent call last):
  File "C:\Users\Q\code\scientizer\main.py", line 14, in <module>
    main()
  File "C:\Users\Q\code\scientizer\main.py", line 11, in main
    retriever.retrieve_papers(query)
  File "C:\Users\Q\code\scientizer\retrievers\mendeley.py", line 21, in retrieve_papers
    authors = ', '.join([author.name for author in paper.authors])
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\Q\code\scientizer\retrievers\mendeley.py", line 21, in <listcomp>
    authors = ', '.join([author.name for author in paper.authors])
                         ^^^^^^^^^^^
  File "C:\Users\Q\code\scientizer\env\Lib\site-packages\mendeley\response.py", line 18, in __getattr__
    raise AttributeError("'%s' object has no attribute '%s'" % (self.__class__.__name__, name))
AttributeError: 'Person' object has no attribute 'name'

ASSISTANT: The error message indicates that the `Person` object does not have a `name` attribute. This means that the way we are trying to access the author's name is incorrect. We need to find the correct attribute or method to access the author's name from the `Person` object provided by the Mendeley API.

Since I don't have access to the actual `mendeley` package documentation or source code, I will assume that the `Person` object has a `first_name` and `last_name` attribute, which is a common way to store a person's name. I will modify the `retrieve_papers` method in the `retrievers\mendeley.py` file to concatenate the `first_name` and `last_name` to get the full name of the author.

Here is the diff for the required change:

```diff
--- retrievers/mendeley.py
+++ retrievers/mendeley.py
@@ ... @@
     def retrieve_papers(self, query):
         papers = self.session.catalog.search(query, view='bib')
         try:
             with Session() as db_session:
                 for paper in papers.iter():
                     # Extract relevant metadata
-                    authors = ', '.join([author.name for author in paper.authors])
+                    authors = ', '.join([f"{author.first_name} {author.last_name}" for author in paper.authors])
                     abstract = paper.abstract
                     altmetric_score = paper.scores.get('altmetric_score')
                     ...
```

This change assumes that the `Person` object has `first_name` and `last_name` attributes. If the attributes are named differently, you will need to adjust the code accordingly. If the `Person` object provides a method to get the full name, you should use that method instead.